+++ packages/shared/src/index.ts
@@ -1,5 +1,6 @@
 // Shared package public API
 export * from './constants'
 export * from './enums'
+export * from './lib'
 export * from './schemas'
 export * from './types'